move RememberMe from workouts -> users
