+++ backend/users/urls.py
@@ -13,4 +13,5 @@ urlpatterns = [
         views.AthleteFileDetail.as_view(),
         name="athletefile-detail",
     ),
+    path("api/remember_me/", views.RememberMe.as_view(), name="remember_me"),
 ]
