+++ backend/workouts/urls.py
@@ -53,6 +53,5 @@ urlpatterns = format_suffix_patterns(
         path("api/auth/", include("rest_framework.urls")),
         path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
         path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
-        path("api/remember_me/", views.RememberMe.as_view(), name="remember_me"),
     ]
 )
